requirements.txt
fastapi==0.115.12
uvicorn==0.34.1
line-bot-sdk==3.16.3
langchain==0.3.24
langchain-community==0.3.23
pydantic==2.11.1
jinja2==3.1.6
notion-client==2.3.0
openai==1.77.0
langchain-openai==0.3.16
langchain-ollama==0.3.2
langchain-google-genai==2.1.4

redis>=5.0
boto3>=1.37.33
python-dotenv>=1.1.0
python-multipart>=0.0.20
requests>=2.32.3
loguru>=0.7.3
httpx>=0.28.1
pillow